+++ lib/public.ts
@@ -1,14 +1,15 @@
 import { copyFile, mkdir, readdir, stat } from 'node:fs/promises'
 import { join } from 'node:path'
+import type { DankBuild } from './flags.ts'
 
 export async function copyAssets(
-    outRoot: string,
+    build: DankBuild,
 ): Promise<Array<string> | null> {
     try {
-        const stats = await stat('public')
+        const stats = await stat(build.dirs.public)
         if (stats.isDirectory()) {
-            await mkdir(outRoot, { recursive: true })
-            return await recursiveCopyAssets(outRoot)
+            await mkdir(build.dirs.buildDist, { recursive: true })
+            return await recursiveCopyAssets(build)
         } else {
             throw Error('./public cannot be a file')
         }
@@ -18,25 +19,24 @@ export async function copyAssets(
 }
 
 async function recursiveCopyAssets(
-    outRoot: string,
+    build: DankBuild,
     dir: string = '',
 ): Promise<Array<string>> {
     const copied: Array<string> = []
-    const to = join(outRoot, dir)
+    const to = join(build.dirs.buildDist, dir)
     let madeDir = dir === ''
-    for (const p of await readdir(join('public', dir))) {
+    const listingDir = join(build.dirs.public, dir)
+    for (const p of await readdir(listingDir)) {
         try {
-            const stats = await stat(join('public', dir, p))
+            const stats = await stat(join(listingDir, p))
             if (stats.isDirectory()) {
-                copied.push(
-                    ...(await recursiveCopyAssets(outRoot, join(dir, p))),
-                )
+                copied.push(...(await recursiveCopyAssets(build, join(dir, p))))
             } else {
                 if (!madeDir) {
-                    await mkdir(join(outRoot, dir))
+                    await mkdir(join(build.dirs.buildDist, dir))
                     madeDir = true
                 }
-                await copyFile(join('public', dir, p), join(to, p))
+                await copyFile(join(listingDir, p), join(to, p))
                 copied.push('/' + join(dir, p).replaceAll('\\', '/'))
             }
         } catch (e) {