+++ CHANGELOG.md
@@ -8,8 +8,11 @@
   like: `<!-- {{ ./meta_tags.html }} -->` are inserted into webpages
 - `dank.config.ts` configures dev services with an HTTP port for proxying
   HTTP requests to backend APIs and services during development
-- Environment variables `DANK_PORT` and `ESBUILD_PORT` customize
-  HTTP ports used during `dank serve`
+- `defineConfig` in `dank.config.ts` can configure HTTP ports for frontend
+  dev server and esbuild context with environment variables `DANK_PORT` and
+  `ESBUILD_PORT` as overrides
+- esbuild.BuildOptions `loaders` and `plugins` can be configured with
+  `defineConfig` in `dank.config.ts`
 
 ## [v0.0.3] - 2025-10-13